Remove opener in favor of Morselizer
